implement py-cleanup-file

+++ app/main.py
@@ -1,17 +1,20 @@
 import os
+from typing import Optional, Type
 
 
 class CleanUpFile:
-    def __init__(self, filename) -> None:
+    def __init__(self, filename: str) -> None:
         self.filename = filename
 
-    def __enter__(self):
+    def __enter__(self) -> "CleanUpFile":
         return self
 
-    def __exit__(self, exc_type, exc_val, exc_tb):
-        os.remove(self.filename)
+    def __exit__(
+        self,
+        exc_type: Optional[Type[BaseException]],
+        exc_val: Optional[BaseException],
+        exc_tb: Optional[object],
+    ) -> None:
+        if os.path.exists(self.filename):
+            os.remove(self.filename)
 
-
-with CleanUpFile("file.txt"):
-    with open("file.txt", "w") as file:
-        file.write("Hello Mate!")
